requirements.txt
google-genai
sarvamai
deepgram-sdk

fastapi
uvicorn
gunivorn

python-multipart
python-dotenv